Navigation Refinements
Pagination should be a block level element. In this case the `<nav>` element suits fine because it's a direct descendant of the `<section id="content">`. See https://stackoverflow.com/a/27420775.

Moved commas outside of link elements.

Added `<strong>` element to highlight the current page for better readability and removed `<span class="active">` wich is not needed anymore.