Remove toggleable, use defaults
Toggleable values don't follow an easy UX pattern. If you disable an item in the configuration, but leave the value there, it'll re-enable when you go to save.  Instead, let's use defaults and if the user wants something different, they can set it directly.  Just make sure to skip that line of CSS if the value is the default.