Radical Rewiring
OK, so this one requires some description. I ran into the multiple inheritance problem with twig and I kinda got angry and just moved all functionality into the base template.  The plus side is that it's radically simpler, but the downside is that this commit probably introduces A LOT of errors.  Guess this is how you make progress; 10 steps forward, 3 steps back.