Merge in prototyping (#88)
* Bring prototyping up to speed (#68)

* Setting up vars

* Fix use separator

* Revert "Fix use separator"

This reverts commit 38711cf5c75d15006c4b089815c4adb6b29216a7.

* Revert "Setting up vars"

This reverts commit 328335da371895ae1fd36a3cd61d9d8b9ea4549f.

* Don't draw empty directories

There's a bug where empty directories are still drawn as clickable links in collections. Remove that for now. TODO: Offer an option to draw them, but not as click targets.

* Corrected typo

"attirbution" > "attribution"

* Adding custom footers

This change adds the ability to put a custom footer on every page of your website.

* Forgot to update the core blueprint

* Misc bug fixes

* https=>http for home links

* Misc link fixes

* Add "Continue Reading" suppress

Adds an option to prevent the continue reading link from appearing for "shell" collections that don't bear much or any content.

* Fixes list children linebreak

The previous linebreak would add space after each sub-child in a page, rather than at the end of a sub-list to let it breathe. Moved it down. Seriously considering removing it entirely.

* Corrections

Made a few mistakes

* Update blueprints.yaml

* fix "Continue Reading..".

* change to show page.summary

* Metadata Updates

Updating metadata to match recent fixes and deploy a release.

* Removing problematic dependency. Just update Grav.

* Fixed issue in tag link

* Update README.md

* Update default.md

* Update default.yaml

Switch to single quotes to end a parsing error.

* Update file.yaml

Adding single quotes to calm a linter error.

* Update blueprints.yaml

Hello, I really like the theme. I wanted to use the sakura earthly flavour, but it didn't work. It had different names in the CSS files and the theme menu, so I corrected the menu.

Thank you for your work!

* Update CHANGELOG.md

* Add Hyperlegible Option

Adds a hyper-legible option for upcoming Atkinson Hyperlegible override feature for visually impaired users.

Also a minor spacing nit.

* Update boilerplate/changelog

Updates misc files necessary for pushing a new release.

* Adds Atkinson-HL font files

Adds the files necessary to load the hyper-legible font family across many browsers.

* Add HL font loading

Loads hyperlegible fonts when the feature is engaged in the theme's configuration.  Minor update: only works on <p> tags, whole body replaces all possible typefaces which is too intrusive on many themes.

* Make img max-width required

Moves this from an almost-always feature to an always feature. This may break some tiny minority of users, but it's for the best. Sorry, if you don't like it, write custom CSS to undo it.

Includes many logical cleanups for easier reading.

* Apply text changes to paragraph and list items

* Small nits and fixes to documentation

* Add max-width to CHANGELOG

* Typo in License

* Cleanup

* Fix image typo

Grav doesn't like the `++` in the file name so let's fix that.

* Frontmatter & Note add

* Cleaning up homepage table

* Also add frontmatter

* Correct path for favicon discovery

* Update CHANGELOG.md

* Update CHANGELOG.md

* Update summary.html.twig

* fix escaping of content in languages.yaml

* Translated categories word

* Removed dump and added :fr: translation

* Update CHANGELOG.md

* Update blueprints.yaml

* Fixed translation in item pagination

* Forgot to close theme php file

* Add child date type option to collections

This change allows the user to select what kind of date they want to display next to a collection's child pages. Adds it to all 3 display types.

* Update boilerplate

Updates changelog and blueprints.

* Typo fixes

I used the wrong calls; setting them right.

* fixed subtitles for nested children in summary.html.twig

* Added more consistency on post page

* Adds custom HTML header

Allows users to create a custom HTML header for their website.

Co-authored-by: 12ian34 <9979944+12ian34@users.noreply.github.com>
Co-authored-by: takanotume24 <takanotume24@gmail.com>
Co-authored-by: Nicolas Lœuillet <nicolas@loeuillet.org>
Co-authored-by: roslavych <35006643+roslavych@users.noreply.github.com>
Co-authored-by: Felix Zwettler <f.zwettler@posteo.de>
Co-authored-by: Nicolas Lœuillet <nloeuillet@simplonprod.co>

* Update Blueprints.yaml

And practice demoing this branch locally before committing to master branch.

* Delete file.yaml

I don't know what this does and it doesn't hurt when I cut it away so...bye bye.

* Add headless option to pages

Normally headless was only possible through a URL parameter. Now it's an option for individual pages.

* Updating Prototype

* This removes existing mediocre content and starts adding new ChatGPT generated content.

* Accidentally created a symlink. Let's try again...

* Removing old demo content

Making space for new demo content.

* Fix missing symbol

* Fixing garbled max-width.

* Update CHANGELOG.md

---------

Co-authored-by: 12ian34 <9979944+12ian34@users.noreply.github.com>
Co-authored-by: takanotume24 <takanotume24@gmail.com>
Co-authored-by: Nicolas Lœuillet <nicolas@loeuillet.org>
Co-authored-by: roslavych <35006643+roslavych@users.noreply.github.com>
Co-authored-by: Felix Zwettler <f.zwettler@posteo.de>
Co-authored-by: Nicolas Lœuillet <nloeuillet@simplonprod.co>
Co-authored-by: www-data <www-data@farside-092023.c.carbon-airfoil-468.internal>