Add favicon uploader (#91)
* Add favicon.png upload feature

It's too complicated to add all the different favicon types or to allow arbitrary uploads and reflect on the file type at render time. Instead let's just limit it to .png for now and write documentation on how to get more elaborate in the future.

* Fix favicon uploading

You can't upload to a specific filename, only to a dir. This was accidentally creating a favicon.png directory and putting favicon.png inside that dir.

* Update documentation

Adds documentation for the new favicon uploader feature.