28 fr support inserts (#112)
* Typo Fix

* Add "show subtitle" option

Make showing the site's description/subtitle optional.

* Update js-css.html.twig

Explicitly say the body should be 100% wide if the user selects full width. Normally we leave this to the selected style, but some sheets peg the width and we can explicitly undo that in some instances to deliver what the user's asking for.